ai-worker/requirements.txt
fastapi
uvicorn[standard]
redis[hiredis]
asyncpg
pydantic
pydantic-settings
httpx
supadata
python-multipart
whisper-openai